fix(inkless): attempt to delete only if error is from control-plane [INK-148]
Currently delete on commit errors is happening regardless of the type of
exception.
Throwing a specific exception on the control-plane implementations to be
properly filtered out to attempt to the deletion.

In the case of the In-memory Control Plane, the synchronization has been
moved from the inner to the top method to ensure a commit happens at a
time, and files are updated at the end.

+++ storage/inkless/src/main/java/io/aiven/inkless/merge/FileMerger.java
@@ -42,6 +42,7 @@ import io.aiven.inkless.config.InklessConfig;
 import io.aiven.inkless.control_plane.BatchInfo;
 import io.aiven.inkless.control_plane.BatchMetadata;
 import io.aiven.inkless.control_plane.ControlPlane;
+import io.aiven.inkless.control_plane.ControlPlaneException;
 import io.aiven.inkless.control_plane.FileMergeWorkItem;
 import io.aiven.inkless.control_plane.MergedFileBatch;
 import io.aiven.inkless.produce.FileUploadJob;
@@ -222,11 +223,14 @@ public class FileMerger implements Runnable {
                     mergedFileBatches
                 );
             } catch (final Exception e) {
-                LOGGER.error("Error committing merged file, attempting to remove the uploaded file {}", objectKey, e);
-                try {
-                    storage.delete(objectKey);
-                } catch (final StorageBackendException e2) {
-                    LOGGER.error("Error removing the uploaded file {}", objectKey, e2);
+                if (e instanceof ControlPlaneException) {
+                    // only attempt to remove the uploaded file if it is a control plane error
+                    LOGGER.error("Error committing merged file, attempting to remove the uploaded file {}", objectKey, e);
+                    try {
+                        storage.delete(objectKey);
+                    } catch (final StorageBackendException e2) {
+                        LOGGER.error("Error removing the uploaded file {}", objectKey, e2);
+                    }
                 }
                 // The original exception will be thrown.
                 throw e;
